Avoid undefined integer operations.
This fixes errors reported by the -fsanitize=undefined sanitizer.

Before accepting the message interval value from a sync message, check
if it is between -10 and 22, same as required for the delay request
interval.

In the calculation of fest/stats/nrate max_count use unsigned 1 to avoid
an invalid shift by 31.

In tmv.h operations cast values to uint64_t to avoid signed overflows
and a left-shift of a negative value.

+++ tmv.h
@@ -49,7 +49,7 @@ typedef struct {
 static inline tmv_t tmv_add(tmv_t a, tmv_t b)
 {
 	tmv_t t;
-	t.ns = a.ns + b.ns;
+	t.ns = (uint64_t)a.ns + (uint64_t)b.ns;
 	return t;
 }
 
@@ -78,7 +78,7 @@ static inline int tmv_is_zero(tmv_t x)
 static inline tmv_t tmv_sub(tmv_t a, tmv_t b)
 {
 	tmv_t t;
-	t.ns = a.ns - b.ns;
+	t.ns = (uint64_t)a.ns - (uint64_t)b.ns;
 	return t;
 }
 
@@ -126,7 +126,7 @@ static inline TimeInterval tmv_to_TimeInterval(tmv_t x)
 	} else if (x.ns > (int64_t)MAX_TMV_TO_TIMEINTERVAL) {
 		return MAX_TMV_TO_TIMEINTERVAL << 16;
 	}
-	return x.ns << 16;
+	return (uint64_t)x.ns << 16;
 }
 
 static inline struct Timestamp tmv_to_Timestamp(tmv_t x)
